chore(DTFS2-7727): split clone-gef-deal into 3 files
